pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archi"
version = "0.1.0"
description = "AI-native architectural and interior design MCP server"
requires-python = ">=3.11"
dependencies = [
    "cadquery-ocp>=7.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
